# docs/03-device-api.openapi.yaml
openapi: 3.0.3
info:
  title: 小卡 Device API
  version: 0.1.0
  description: |
    设备端（ESP32-S3）API 规范 v1 —— 权威文档见 `02-device-api-v1.md`。
    本文件用于导入 Postman/Apifox 调试或生成 mock server。
    Web 社区端接口（注册/登录/好友请求/绑定设备等）不在本文件范围内。
servers:
  - url: https://api-dev.example.com/v1
    description: dev 环境（联调用）
  - url: https://api.example.com/v1
    description: 生产
...
paths:
  /pair/code:
    post:
      tags: [pairing]
      summary: 领取配对码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_id]
              properties:
                device_id: { type: string, example: dvc_a1b2c3d4e5f6 }
                fw_version: { type: string, example: "0.1.0" }
                hw: { type: string, example: cams3-lite }
      responses:
        "200":
          description: 配对码
          content:
            application/json:
              schema:
                type: object
                properties:
                  pair_code: { type: string, example: "482913" }
                  expires_in: { type: integer, example: 600 }
  /pair/status:
    get:
      tags: [pairing]
      summary: 轮询配对状态
      security: []
      parameters:
        - { name: device_id, in: query, required: true, schema: { type: string } }
        - { name: pair_code, in: query, required: true, schema: { type: string } }
      responses:
        "202":
          description: 等待用户绑定
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: pending }
        "200":
          description: 绑定成功，返回设备 token
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: bound }
                  device_token: { type: string, example: ak_9f2e64chars }
                  user:
                    type: object
                    properties:
                      username: { type: string }
                      display_name: { type: string }
        "410":
          description: 配对码过期
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /photos/{photo_id}/image:
    get:
      tags: [photos]
      summary: 下载照片（320x240 JPEG）
      parameters:
        - { name: photo_id, in: path, required: true, schema: { type: string } }
        - { name: size, in: query, required: false, schema: { type: integer, default: 320 }, description: "预留，MVP 仅一档" }
      responses:
        "200":
          description: "JPEG 字节流（需认证且通过好友可见性校验，Cache-Control: private immutable）"
          content:
            image/jpeg:
              schema: { type: string, format: binary }
        "404":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: 无权查看该照片
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/StorageUnavailable" }
...
  /device/config:
    post:
      tags: [device]
      summary: Web 为已绑定设备下发玩法配置
      description: 由设备主人在 Web 端调用；设备下一次读取 `/device/state` 时取得 pending_config，回执后转为 active_config。
      parameters:
        - { name: Idempotency-Key, in: header, required: false, schema: { type: string }, description: 同一设备重试同 key 不会重复排队 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_id, filter_id, play_type]
              properties:
                device_id: { type: string }
                filter_id: { type: string, enum: [none, warm, bw, film, vivid] }
                play_type: { type: string, enum: [beauty, ccd, template] }
                beauty: { type: integer, minimum: 0, maximum: 100, default: 0 }
                sticker: { type: string, enum: [none, star, date], default: none }
      responses:
        "202":
          description: 配置已排队
          content:
            application/json:
              schema:
                type: object
                properties:
                  config_id: { type: string }
                  status: { type: string, example: queued }
                  device_id: { type: string }
                  config: { $ref: "#/components/schemas/DeviceConfig" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /device/heartbeat:
    post:
      tags: [device]
      summary: 心跳（可选，调试用）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                battery: { type: integer }
                rssi: { type: integer }
                fw_version: { type: string }
                free_heap: { type: integer }
      responses:
        "204": { description: 已接收 }
  /device/ack:
    post:
      tags: [device]
      summary: 回执轻信号
      description: 设备确认已处理下发的动态或反应；需要 device_token。
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                config_id: { type: string, description: 已处理的玩法配置 ID，可选 }
      responses:
        "204": { description: 已接收 }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "device_token，配对成功后获得"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: PHOTO_TOO_LARGE }
            message: { type: string }
...
    DeviceConfig:
      type: object
      required: [id, filter_id, play_type, beauty, sticker, updated_at]
      properties:
        id: { type: string, example: cfg_01J8XYZ }
        filter_id: { type: string, enum: [none, warm, bw, film, vivid] }
        play_type: { type: string, enum: [beauty, ccd, template] }
        beauty: { type: integer, minimum: 0, maximum: 100 }
        sticker: { type: string, enum: [none, star, date] }
        updated_at: { type: string, format: date-time }
  responses:
    Unauthorized:
      description: "token 无效/被解绑 → 清 NVS token，回配对流程"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    StorageUnavailable:
      description: "图片存储暂不可用；请求未完成，稍后重试"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: 请求参数不合法
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: 当前账号无权操作该设备
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: 设备不存在或尚未绑定
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
security:
  - bearerAuth: []